Implement PR feedback
