+++ flutter_google_places_sdk_platform_interface/lib/src/types/author_attribution.dart
@@ -4,7 +4,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'author_attribution.freezed.dart';
 part 'author_attribution.g.dart';
 
-@freezed
+@Freezed()
 class AuthorAttribution with _$AuthorAttribution {
   /// Constructs a [AuthorAttribution] object.
   const factory AuthorAttribution({
