+++ flutter_google_places_sdk_platform_interface/lib/src/types/lat_lng_bounds.dart
@@ -22,17 +22,3 @@ class LatLngBounds with _$LatLngBounds {
   factory LatLngBounds.fromJson(Map<String, Object?> json) =>
       _$LatLngBoundsFromJson(json);
 }
-
-/// An immutable class representing a latitude/longitude aligned circle, with a defined radius.
-@Freezed()
-class CircularBounds with _$CircularBounds {
-  /// constructs a [CircularBounds] object.
-  const factory CircularBounds({
-    required LatLng center,
-    required double radius,
-  }) = _CircularBounds;
-
-  /// Parse an [CircularBounds] from json.
-  factory CircularBounds.fromJson(Map<String, Object?> json) =>
-      _$CircularBoundsFromJson(json);
-}
