+++ flutter_google_places_sdk_platform_interface/lib/src/types/place_type_filter.dart
@@ -57,6 +57,5 @@ enum PlaceTypeFilter {
 }
 
 extension PlaceTypeFilterDescriptor on PlaceTypeFilter {
-  @Deprecated('Use `name` instead')
-  String get value => describeEnum(this);
+  String get value => name;
 }
