+++ flutter_google_places_sdk_platform_interface/pubspec.yaml
@@ -11,7 +11,7 @@ dependencies:
     sdk: flutter
   freezed_annotation: ^2.4.1
   json_annotation: ^4.8.1
-  plugin_platform_interface: ^2.1.8
+  plugin_platform_interface: ^2.1.6
 
 dev_dependencies:
   build_runner: ^2.4.6
