+++ flutter_google_places_sdk_platform_interface/test/method_channel_flutter_google_places_sdk_test.dart
@@ -260,6 +260,8 @@ class FlutterGooglePlacesSdkPlatformMock extends Mock
     with MockPlatformInterfaceMixin
     implements FlutterGooglePlacesSdkPlatform {}
 
-class ImplementsFlutterGooglePlacesSdkPlatform extends Mock implements FlutterGooglePlacesSdkPlatform {}
+class ImplementsFlutterGooglePlacesSdkPlatform extends Mock
+    implements FlutterGooglePlacesSdkPlatform {}
 
-class ExtendsFlutterGooglePlacesSdkPlatform extends FlutterGooglePlacesSdkPlatform {}
+class ExtendsFlutterGooglePlacesSdkPlatform
+    extends FlutterGooglePlacesSdkPlatform {}
